modularize tests for qchem inputs
